REF: move roperators to pandas.core (#40444)
