+++ pandas/core/internals/blocks.py
@@ -103,6 +103,7 @@ from pandas.core.arrays import (
 )
 from pandas.core.base import PandasObject
 import pandas.core.common as com
+import pandas.core.computation.expressions as expressions
 from pandas.core.construction import (
     ensure_wrapped_if_datetimelike,
     extract_array,
@@ -1307,8 +1308,6 @@ class Block(PandasObject):
         -------
         List[Block]
         """
-        import pandas.core.computation.expressions as expressions
-
         assert cond.ndim == self.ndim
         assert not isinstance(other, (ABCIndex, ABCSeries, ABCDataFrame))
