+++ pandas/core/ops/methods.py
@@ -8,16 +8,7 @@ from pandas.core.dtypes.generic import (
     ABCSeries,
 )
 
-from pandas.core.ops.roperator import (
-    radd,
-    rdivmod,
-    rfloordiv,
-    rmod,
-    rmul,
-    rpow,
-    rsub,
-    rtruediv,
-)
+from pandas.core.ops import roperator
 
 
 def _get_method_wrappers(cls):
@@ -89,19 +80,19 @@ def _create_methods(cls, arith_method, comp_method):
     new_methods.update(
         {
             "add": arith_method(operator.add),
-            "radd": arith_method(radd),
+            "radd": arith_method(roperator.radd),
             "sub": arith_method(operator.sub),
             "mul": arith_method(operator.mul),
             "truediv": arith_method(operator.truediv),
             "floordiv": arith_method(operator.floordiv),
             "mod": arith_method(operator.mod),
             "pow": arith_method(operator.pow),
-            "rmul": arith_method(rmul),
-            "rsub": arith_method(rsub),
-            "rtruediv": arith_method(rtruediv),
-            "rfloordiv": arith_method(rfloordiv),
-            "rpow": arith_method(rpow),
-            "rmod": arith_method(rmod),
+            "rmul": arith_method(roperator.rmul),
+            "rsub": arith_method(roperator.rsub),
+            "rtruediv": arith_method(roperator.rtruediv),
+            "rfloordiv": arith_method(roperator.rfloordiv),
+            "rpow": arith_method(roperator.rpow),
+            "rmod": arith_method(roperator.rmod),
         }
     )
     new_methods["div"] = new_methods["truediv"]
@@ -109,7 +100,7 @@ def _create_methods(cls, arith_method, comp_method):
     if have_divmod:
         # divmod doesn't have an op that is supported by numexpr
         new_methods["divmod"] = arith_method(divmod)
-        new_methods["rdivmod"] = arith_method(rdivmod)
+        new_methods["rdivmod"] = arith_method(roperator.rdivmod)
 
     new_methods.update(
         {
