+++ pandas/core/roperator.py
